Refactors Project Structure
