+++ src/main/java/com/hideyoshi/backendportfolio/base/auth/model/RoleToUserDTO.java
@@ -1,0 +1,10 @@
+package com.hideyoshi.backendportfolio.base.auth.model;
+
+import lombok.Data;
+
+@Data
+public
+class RoleToUserDTO {
+    private String username;
+    private String roleName;
+}
