+++ src/main/java/com/hideyoshi/backendportfolio/base/auth/oauth/mapper/OAuthMap.java
@@ -1,0 +1,13 @@
+package com.hideyoshi.backendportfolio.base.auth.oauth.mapper;
+
+import com.hideyoshi.backendportfolio.base.auth.entity.Provider;
+
+public interface OAuthMap {
+
+    String getPrincipal();
+
+    String getProfilePicture();
+
+    Provider getProvider();
+
+}
