+++ src/main/java/com/hideyoshi/backendportfolio/base/auth/repo/UserRepository.java
@@ -1,0 +1,14 @@
+package com.hideyoshi.backendportfolio.base.auth.repo;
+
+import com.hideyoshi.backendportfolio.base.auth.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findByUsername(String username);
+
+    Optional<User> findByEmail(String email);
+}
