--- src/main/java/com/hideyoshi/backendportfolio/base/security/oauth/mapper/OAuthMap.java
@@ -1,13 +1,0 @@
-package com.hideyoshi.backendportfolio.base.security.oauth.mapper;
-
-import com.hideyoshi.backendportfolio.base.user.entity.Provider;
-
-public interface OAuthMap {
-
-    String getPrincipal();
-
-    String getProfilePicture();
-
-    Provider getProvider();
-
-}
