+++ src/main/java/com/hideyoshi/backendportfolio/base/session/api/SessionController.java
@@ -1,6 +1,6 @@
 package com.hideyoshi.backendportfolio.base.session.api;
 
-import com.hideyoshi.backendportfolio.base.security.model.AuthDTO;
+import com.hideyoshi.backendportfolio.base.auth.model.AuthDTO;
 import com.hideyoshi.backendportfolio.base.session.service.SessionManagerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
