--- src/main/java/com/hideyoshi/backendportfolio/base/user/model/RoleToUserDTO.java
@@ -1,10 +1,0 @@
-package com.hideyoshi.backendportfolio.base.user.model;
-
-import lombok.Data;
-
-@Data
-public
-class RoleToUserDTO {
-    private String username;
-    private String roleName;
-}
